+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - models are now stored on `req.curdy` instead of just the `req`
 - Arrays are now marked as dirty
   - As a result updating an array of subdocuments behaves differently than an array of strings
-- Wraps api model data in the model name
+- Wraps api model render data in the model name
   - Show, Create, Delete, and Update use the singular name
   - ShowAll uses the plural name